pyproject.toml
[tool.poetry]
authors = ["Naveen Thurimerla <nawinto99@gmail.com>"]
description = "The Black Ox is a database migration assistant tool, this tool is used to check for compatibility issues as well as the database's performance."
name = "black-ox"
version = "0.1.0"

[tool.poetry.dependencies]
GitPython = "^3.1.24"
PyYAML = "^6.0.0"
mkdocs = {version = "^1.2.3", optional = true}
mkdocs-autorefs = {version = "^0.3.0", optional = true}
mkdocs-include-markdown-plugin = {version = "^3.2.3", optional = true}
mkdocs-material = {version = "^7.3.4", optional = true}
mkdocs-material-extensions = {version = "^1.0.3", optional = true}
mkdocstrings = {version = "^0.16.2", optional = true}
psycopg2 = "^3.0.1"
python = "^3.9"

[tool.poetry.extras]
doc = ["mkdocs", "mkdocs-include-markdown-plugin", "mkdocs-material", "mkdocstrings", "mkdocs-material-extensions", "mkdocs-autorefs"]

[tool.poetry.dev-dependencies]
pytest = "^6.2.5"

[build-system]
build-backend = "poetry.core.masonry.api"
requires = ["poetry-core>=1.0.0"]
